modal component stories and file renamed
