+++ packages/frontend-app/src/app/components/Modal/Modal.scss
