+++ packages/frontend-app/src/app/components/Modal/Modal.tsx
@@ -4,7 +4,7 @@ import { useState, useRef, useEffect } from "react";
 import { createPortal } from "react-dom";
 
 import { sleep } from "../../../utils/sleep";
-import "./index.scss";
+import "./Modal.scss";
 import useMediaQuery, { screens } from "../../hooks/useMediaQuery";
 import useSsr from "../../hooks/useSsr";
 import { Button } from "../Button/Button";
